Feature/3

+++ src/views/Diagnosis/index.js
@@ -36,7 +36,12 @@ function Diagnosis (props) {
 
     const selectPatient = (child_patient_id, child_patient_name, child_patient_state) => {
         if(child_patient_state !== "진료 중"){
-            alert("진료 상태인 환자만 선택 가능합니다.");
+            Swal.fire({
+                icon: 'info',
+                title: '진료 상태인 환자만 선택 가능합니다.',
+                showConfirmButton: false,
+                timer: 1500
+            })
         }
         else{
             setSelectP({
@@ -105,6 +110,14 @@ function Diagnosis (props) {
     };
 
     const deleteAll = () => {  //전체 삭제
+        if(selectSymptoms.length === 0){
+            Swal.fire({
+                icon: 'info',
+                title: '삭제 할 증상이 없습니다.',
+                showConfirmButton: false,
+                timer: 1500
+            })
+        }
         const symptomSelect = selectSymptoms.filter(symptom => symptom.search_id !== symptom.search_id);
         setSelectSymptoms([
             ...symptomSelect
@@ -144,67 +157,80 @@ function Diagnosis (props) {
     const [testFlag, setTestFlag] = useState(1);
 
     const testRequest = (event) => { //검사 요청
-        console.log(event.patient_id)
-
-        let flag = true;
-        for(let op of opinions) {
-            if(op.diagnostic_test_state === '검사 중') {
-                if(op.patient_id === event.patient_id)
-                    flag = false;
-            }
-        }
-
-        if(selectedPatient.patient_id && flag && selectSymptoms.length !== 0) {
-
-            let time = new Date();
-            let yy = time.getFullYear();
-            let tempmm = time.getMonth() + 1;
-            let dd = time.getDate();
-
-            if(10 > tempmm){
-                let mm = '0'+tempmm;
-                let myDateString = yy + '-' + mm + '-' + dd;
-                setOpinions(opinions.concat({
-                    receipt_id: newReceipt_id,
-                    receipt_opinion: '검사 완료 후 소견 작성 필요',
-                    receipt_uniqueness: '',
-                    receipt_datetime: myDateString,
-                    patient_id: selectedPatient.patient_id+'',
-                    patient_name: selectedPatient.patient_name+'',
-                    diagnostic_test_state:'검사 중',
-                    medicines: [],
-                    test_flag: testFlag
-                }));
-                setNewReceipt_id(newReceipt_id + 1);
-                setTestFlag(testFlag +1);
-                console.log(opinions)
-            }
-            else{
-                let myDateString = yy + '-' + tempmm + '-' + dd;
-                setOpinions(opinions.concat({
-                    receipt_id: newReceipt_id,
-                    receipt_opinion: '검사 완료 후 작성 필요',
-                    receipt_uniqueness: '',
-                    receipt_datetime: myDateString,
-                    patient_id: selectedPatient.patient_id+'',
-                    patient_name: selectedPatient.patient_name+'',
-                    diagnostic_test_state:'검사 중',
-                    medicines: [],
-                    test_flag: testFlag
-                }));
-                setNewReceipt_id(newReceipt_id + 1);
-                setTestFlag(testFlag +1);
-            }
-            
-            deleteAll();
-        }
+       
         
+
+                let flag = true;
+                for(let op of opinions) {
+                    if(op.diagnostic_test_state === '검사 중') {
+                        if(op.patient_id === event.patient_id){
+                            flag = false;
+                            Swal.fire({
+                                icon: 'error',
+                                title: '이미 검사 중인 목록이 있습니다.',
+                                showConfirmButton: false,
+                                timer: 1500
+                            })
+                        }
+                    }
+                }
+
+                if(selectedPatient.patient_id && flag && selectSymptoms.length !== 0) {
+
+                    let time = new Date();
+                    let yy = time.getFullYear();
+                    let tempmm = time.getMonth() + 1;
+                    let dd = time.getDate();
+
+                    if(10 > tempmm){
+                        let mm = '0'+tempmm;
+                        let myDateString = yy + '-' + mm + '-' + dd;
+                        setOpinions(opinions.concat({
+                            receipt_id: newReceipt_id,
+                            receipt_opinion: '검사 완료 후 소견 작성 필요',
+                            receipt_uniqueness: '',
+                            receipt_datetime: myDateString,
+                            patient_id: selectedPatient.patient_id+'',
+                            patient_name: selectedPatient.patient_name+'',
+                            diagnostic_test_state:'검사 중',
+                            medicines: [],
+                            test_flag: testFlag
+                        }));
+                        setNewReceipt_id(newReceipt_id + 1);
+                        setTestFlag(testFlag +1);
+                        console.log(opinions)
+                    }
+                    else{
+                        let myDateString = yy + '-' + tempmm + '-' + dd;
+                        setOpinions(opinions.concat({
+                            receipt_id: newReceipt_id,
+                            receipt_opinion: '검사 완료 후 작성 필요',
+                            receipt_uniqueness: '',
+                            receipt_datetime: myDateString,
+                            patient_id: selectedPatient.patient_id+'',
+                            patient_name: selectedPatient.patient_name+'',
+                            diagnostic_test_state:'검사 중',
+                            medicines: [],
+                            test_flag: testFlag
+                        }));
+                        setNewReceipt_id(newReceipt_id + 1);
+                        setTestFlag(testFlag +1);
+                    }
+                    
+                    deleteAll();
+                    Swal.fire({
+                        icon: 'success',
+                        title: '검사 요청을 완료하였습니다.',
+                        showConfirmButton: false,
+                        timer: 1500
+                    })
+                }
+               
     }; 
      
     /*약품 관련 */
     const [isModalVisible, setIsModalVisible] = useState(false);
     const [medicines, setMedicines] =  useState([]);
-    const [copymedic, setCopyMedic] = useState([]);
     
     const handleModal = () => { //모달 창 열기, 닫기      
         setIsModalVisible(!isModalVisible);
@@ -259,8 +285,18 @@ function Diagnosis (props) {
     
 
     const deleteMedicineAll = () => {  //전체 삭제
-        const deleteAll = medicines && medicines.filter(medicine => medicine.medicine_id !== medicine.medicine_id);
-        setMedicines(deleteAll); 
+        if(medicines.length === 0){
+            Swal.fire({
+                icon: 'info',
+                title: '추가된 약이 없습니다.',
+                showConfirmButton: false,
+                timer: 1500
+            })
+        }
+        else{
+            const deleteAll = medicines && medicines.filter(medicine => medicine.medicine_id !== medicine.medicine_id);
+            setMedicines(deleteAll); 
+        }
     };
     const deleteMedicine = (event) => {  // 삭제
         const deleteAll = medicines.filter(medicine => medicine.medicine_id !== event);
@@ -449,14 +485,19 @@ function Diagnosis (props) {
     const saveOpinion = (event) => {
         
         
-        for(let i of opinions){
+    for(let i of opinions){
             
-            if(event === i.receipt_id){
-              i.receipt_opinion = opp.receipt_opinion;
-              i.receipt_uniqueness = opp.receipt_uniqueness;
-            }
-      }
-      alert("수정이 완료되었습니다.")
+        if(event === i.receipt_id){
+            i.receipt_opinion = opp.receipt_opinion;
+            i.receipt_uniqueness = opp.receipt_uniqueness;
+        }
+    }
+    Swal.fire({
+        icon: 'success',
+        title: '수정이 완료되었습니다.',
+        showConfirmButton: false,
+        timer: 1500
+    })
       closeUpdateModal();
       deleteMedicineAll();  
     }
@@ -470,8 +511,12 @@ function Diagnosis (props) {
             }
             
       }
-    
-       alert("약 처방이 완료되었습니다.");
+       Swal.fire({
+        icon: 'success',
+        title: '약 처방이 완료되었습니다.',
+        showConfirmButton: false,
+        timer: 1500
+    })
        
        setMedicines(medicines && medicines.filter(medicine => medicine.medicine_id !== medicine.medicine_id)); 
     }
@@ -490,11 +535,6 @@ function Diagnosis (props) {
             receipt_id: event1,
             diagnostic_test_state: event2
         });
-        
-
-
-        
-
     }
     useEffect(() => {
         console.log("증상 선택 및 소견 추가시 재실행")
